requirements.txt
# Install requirements for GiftStarer Client

lxml
GoogleAppEngineCloudStorageClient
oauthlib
stripe
requests_oauthlib
mock
tl.testing
